Implemented Fixed Update and DeltaTime

+++ Source/Engines/Scene.cpp
@@ -18,7 +18,7 @@ void Scene::Start()
 	}
 }
 
-void Scene::Update(float delta)
+void Scene::Update(double delta)
 {
 	for (auto actor : actors)
 	{
