+++ Source/Engines/Scene.h
@@ -10,5 +10,5 @@ private:
 public:
 	void Initialize();
 	void Start();
-	void Update(float delta);
+	void Update(double delta);
 };
